changed rating type to float in item and main to fix accuracy of average rating

+++ item/src/main/java/com/example/item/Review.java
@@ -4,13 +4,13 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class Review {
     @JsonProperty("rating")
-    private Long rating;
+    private float rating;
 
     public Review(Long rating) {
         this.rating = rating;
     }
 
-    public Long getRating() {
+    public float getRating() {
         return rating;
     }
